+++ libs/jacdac/bridge.ts
@@ -1,0 +1,15 @@
+namespace jacdac {
+    export class BridgeDriver extends Driver {
+        constructor(name: string) {
+            super(name, 0, DAL.JD_DRIVER_CLASS_BRIDGE);
+            this.supressLog = true; // too verbose
+        }
+
+        start() {
+            if (!this.hasProxy()) {
+                super.start();
+                if (this._proxy) this._proxy.setBridge();
+            }
+        }
+    }
+}
