+++ libs/jacdac/pxt.json
@@ -15,6 +15,7 @@
         "actuator.ts",
         "messagebus.ts",
         "console.ts",
+        "bridge.ts",
         "ns.ts"
     ],
     "testFiles": [
